Use size_t type in rendering interfaces

+++ RealEngine/rendering/batches/GeometryBatch.cpp
@@ -60,17 +60,17 @@ void GeometryBatch::begin() {
 
 void GeometryBatch::end() {
 	//Uploading to VBO
-	int totalSize = 0;
+	size_t totalSize = 0;
 	for (auto& vector : m_vertices) {
-		totalSize += static_cast<int>(vector.size());
+		totalSize += vector.size();
 	}
-	int offset = 0;
+	size_t offset = 0;
 
 	m_buf.redefine(totalSize * sizeof(VertexPOCO), nullptr);
 	for (size_t i = 0u; i < PRIMITIVES_COUNT + SHAPES_COUNT; ++i) {
 		if (m_vertices[i].empty()) continue;
 		m_buf.overwrite(offset, m_vertices[i]);
-		offset += static_cast<int>(m_vertices[i].size() * sizeof(VertexPOCO));
+		offset += m_vertices[i].size() * sizeof(VertexPOCO);
 	}
 }
 
@@ -78,12 +78,12 @@ void GeometryBatch::draw() {
 	m_shaderProgram->use();
 	m_va.bind();
 
-	size_t offset = 0u;
+	int offset = 0u;
 
 	for (size_t i = 0u; i < PRIMITIVES_COUNT + SHAPES_COUNT; ++i) {
 		if (m_vertices[i].empty()) continue;
-		m_va.renderElementsBaseVertex(convertPrimEnum(i), static_cast<int>(m_indices[i].size()), IndexType::UNSIGNED_INT, m_indices[i].data(), static_cast<int>(offset));
-		offset += m_vertices[i].size();
+		m_va.renderElementsBaseVertex(convertPrimEnum(i), m_indices[i].size(), IndexType::UNSIGNED_INT, m_indices[i].data(), offset);
+		offset += static_cast<int>(m_vertices[i].size());
 	}
 
 	m_va.unbind();
